+++ lib/react/test/utils.rb
@@ -4,12 +4,12 @@ module React
       `var ReactTestUtils = React.addons.TestUtils`
 
       def self.render_into_document(element, options = {})
-        raise "You should pass a valid React::Element" unless React.is_valid_element(element)
+        raise "You should pass a valid React::Element" unless React.is_valid_element?(element)
         native_instance = `ReactTestUtils.renderIntoDocument(#{element.to_n})`
 
         if `#{native_instance}._getOpalInstance !== undefined`
           `#{native_instance}._getOpalInstance()`
-        elsif `ReactTestUtils.isDOMComponent(#{native_instance}) && React.findDOMNode !== undefined` 
+        elsif `ReactTestUtils.isDOMComponent(#{native_instance}) && React.findDOMNode !== undefined`
           `React.findDOMNode(#{native_instance})`
         else
           native_instance